backend/pyproject.toml
[tool.poetry]
name = "carbon-footprint-tracker-backend"
version = "0.1.0"
description = "Backend for Carbon Footprint Tracker Lite"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
email-validator = "^2.1.0"
celery = "^5.3.4"
redis = "^5.0.1"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"